Generic Collectible, Platform, and Texture Rework
Created a sprite (and prefab) for a generic collectible and generic platform. Textures were also reworked so that each texture is its own image file, which will eventually be compiled into a texture atlas. 3D assets and prefabs were edited to reflect the change in how the textures are organized.