FingerGunMan Omnidirectional Shooting Rework
Adjusted FingerGunMan rig to allow easy 360 degree shooting. New prefab is "FingerGunManSeparateHand", and should be prototype-ready, minus troubleshooting for the animations.
NOTE: There is a problem with either the Blender export or the Unity import when there tracks in an action that belong to non-existent bones. The file shows as having absolutely no animation.