Generic Collectible Animations
Added animations for the generic collectible prefab. Animations include an idle animation that loads as the entry state, an animation for when the collectible is picked up (particle system should be triggered, via script, when this animation is triggered), and a rotation animation to be used at the developer's discretion as flavor.
Also includes an animation controller with triggers.