Character Shaders
Split the unified character shader into a shader for enemies, and a shader for the main character. Implemented materials for all characters in the game so far, including the elemental.
Split was necessary due to how alpha is calculated at runtime: one shader is targeted at characters that have self occlusion issues (elemental, for example), and the other is targeted at characters that need semitransparency (FGM).