DoozyUI Setup
Imported packages and set up DoozyUI. Part of this testing involves getting rid of our Game Event system to prevent interference with Doozy's Game Event system. If we use Doozy on a project in the future, we will need combine these systems, or remove ours entirely.
THIS BRANCH IS NOT INTENDED FOR USE.